+++ src/Network/Haskoin/Store/CacheWriter.hs
@@ -53,7 +53,7 @@ import           Network.Haskoin.Store.Common           (Balance (..),
                                                          XPubBal (..),
                                                          XPubSpec (..),
                                                          XPubUnspent (..),
-                                                         sortTxs,
+                                                         nullBalance, sortTxs,
                                                          xPubAddrFunction,
                                                          xPubBals, xPubTxs,
                                                          xPubUnspents)
@@ -147,7 +147,8 @@ newXPubC xpub = do
     x <- asks cacheWriterMin
     when empty $ do
         bals <- lift $ xPubBals xpub
-        when (x <= length bals) (go bals)
+        let l = length $ filter (not . nullBalance . xPubBal) bals
+        when (x <= l) (go bals)
   where
     go bals = do
         utxo <- lift $ xPubUnspents xpub Nothing 0 Nothing